30th commit Changes: 1) Added custom messages for incorrect type of user input over-writing the Spring default messages that were too long and not user-friendly. 2) Implement balance change on accounts if transaction was deleted 3) Added checking if account already exist when new account is created. 4) Minor changes.

+++ transaction-service/src/main/java/com/bank/transactionservice/service/TransactionServiceImpl.java
@@ -64,8 +64,13 @@ public class TransactionServiceImpl implements TransactionService {
     @Override
     @Transactional
     public void deleteTransaction(Long transactionId) {
+        Transaction transaction = repository.findById(transactionId).orElse(null);
+        if (transaction == null) {
+            // TODO return message to the web-service
+            throw new EntityNotFoundException("Transaction with id " + transactionId + " not found");
+        }
         repository.deleteById(transactionId);
-        publisher.publishTransactionDeletedEvent(transactionId);
+        publisher.publishTransactionDeletedEvent(transaction);
         log.info("Transaction with id: {} deleted", transactionId);
     }
